swapped out karma for jest
now using yarn

adding typescript tests for decorators
