+++ .babelrc
@@ -1,0 +1,4 @@
+{
+  "presets": ["es2015", "stage-0"],
+  "plugins": ["transform-runtime", "transform-decorators-legacy", "transform-flow-strip-types"]
+}
