+++ .npmignore
@@ -17,3 +17,6 @@ NOTICE
 webpack.dev.config.js
 webpack.prod.config.js
 esp-js-*.tgz
+.babelrc
+jest.config
+yarn.lock
