+++ jest.config
@@ -1,0 +1,12 @@
+{
+    "setupFiles": [
+        "<rootDir>/tests/__jest__/configureEspLogging.js"
+    ],
+    "testRegex": "/tests/.*\\Tests\\.(js|ts)$",
+    "testResultsProcessor": "node_modules/jest-teamcity-reporter",
+    "moduleFileExtensions": [
+        "ts",
+        "tsx",
+        "js"
+    ]
+}
