+++ package.json
@@ -13,9 +13,9 @@
     "event router"
   ],
   "scripts": {
-    "start": "npm run test",
-    "test": "karma start",
-    "test:ci": "karma start --single-run",
+    "start": "yarn test",
+    "test": "jest --watchAll --no-cache --config jest.config",
+    "test:ci": "jest --teamcity --config jest.config --verbose",
     "build": "npm run build:dev && npm run build:prod",
     "build:dev": "webpack --config webpack.dev.config.js",
     "build:prod": "webpack -p --config webpack.prod.config.js"
@@ -27,10 +27,12 @@
   "typings": "./esp.d.ts",
   "devDependencies": {
     "babel-core": "^6.4.5",
+    "babel-jest": "^18.0.0",
     "babel-loader": "^6.2.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-flow-strip-types": "^6.5.0",
     "babel-plugin-transform-runtime": "^6.4.3",
+    "babel-polyfill": "^6.20.0",
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-react": "^6.3.13",
     "babel-preset-stage-0": "^6.3.13",
@@ -39,12 +41,10 @@
     "eslint-loader": "^1.2.1",
     "inject-loader": "^2.0.1",
     "jasmine-core": "^2.3.4",
-    "karma": "^0.13.19",
-    "karma-chrome-launcher": "^0.2.2",
-    "karma-jasmine": "^0.3.7",
-    "karma-spec-reporter": "0.0.23",
-    "karma-webpack": "^1.7.0",
+    "jest": "17.0.3",
+    "jest-teamcity-reporter": "0.2.0",
     "source-map-loader": "^0.1.5",
+    "typescript": "^2.1.4",
     "webpack": "^1.12.12",
     "webpack-dev-server": "^1.14.1"
   }
