+++ webpack.dev.config.js
@@ -39,11 +39,7 @@ module.exports = {
                 include: [
                     path.resolve(__dirname, "src"),
                 ],
-                test: /\.js$/,
-                query: {
-                    presets: ['es2015', 'stage-0'],
-                    plugins: ['transform-runtime', 'transform-decorators-legacy', 'transform-flow-strip-types']
-                }
+                test: /\.js$/
             }
         ]
     },
