+++ webpack.prod.config.js
@@ -37,11 +37,7 @@ module.exports = {
                 include: [
                     path.resolve(__dirname, "src"),
                 ],
-                test: /\.js$/,
-                query: {
-                    presets: ['es2015', 'stage-0'],
-                    plugins: ['transform-runtime', 'transform-decorators-legacy', 'transform-flow-strip-types']
-                }
+                test: /\.js$/
             }
         ]
     },
